Add `keep_disabled_in_children!`
When using Pitchfork, `keep_disabled` isn't quite enough
because `fork_sibling` use a double fork like daemonization,
but also because you likely want GRPC to remain paused in
the mold.

One thing to note though is that the `grpc` maintainer explicitly
said `postfork_child` MUST be called in the immediate child,
but I don't see why that is, so worth a try.